src/Cameras/FaceRecognition/requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
pyserial==3.5
